refactor: action script
