+++ scripts/utils/isValidJson.js
@@ -1,0 +1,9 @@
+module.exports = (jsonString) => {
+    try {
+      JSON.parse(jsonString);
+      return true;
+    } catch (error) {
+      return false;
+    }
+  }
+
